feat: support multiple juice_mux_listen callbacks
Supports multiple callbacks for unhandled STUN requests.

Stores multiple registries, one per local listening port in a
`conn_registry_t **registries;` field on `conn_mode_entry_t`.

The list of registries grows in the same way as the list of
agents in `conn_mode_entry_t`.

Adds a `port` property to `conn_registry_t` which is the port
the registry is bound to, and an `index` property which is the
index of the registry in the containing `registries` list in
`conn_mode_entry_t`.